Moving getBrand() into the plugin.
