+++ express-zod-api/src/metadata.ts
@@ -1,14 +1,2 @@
-import type { z } from "zod";
-
+// @todo consider removal
 export const metaSymbol = Symbol.for("express-zod-api");
-
-export const getBrand = (subject: z.core.$ZodType) => {
-  const { brand } = subject._zod.bag || {};
-  if (
-    typeof brand === "symbol" ||
-    typeof brand === "string" ||
-    typeof brand === "number"
-  )
-    return brand;
-  return undefined;
-};
