+++ express-zod-api/tests/form-schema.spec.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { ez } from "../src";
 import { ezFormBrand } from "../src/form-schema";
-import { getBrand } from "../src/metadata";
+import { getBrand } from "@express-zod-api/zod-plugin";
 
 describe("ez.form()", () => {
   describe("creation", () => {
