+++ zod-plugin/helpers.ts
@@ -1,0 +1,12 @@
+import type { z } from "zod";
+
+export const getBrand = (subject: z.core.$ZodType) => {
+  const { brand } = subject._zod.bag || {};
+  if (
+    typeof brand === "symbol" ||
+    typeof brand === "string" ||
+    typeof brand === "number"
+  )
+    return brand;
+  return undefined;
+};
