+++ zod-plugin/index.ts
@@ -1,1 +1,2 @@
 import "./augmentation"; // side effects here
+export { getBrand } from "./helpers";
